refactor:  moved index to ./src
