+++ package.json
@@ -2,7 +2,7 @@
   "name": "ws_task",
   "version": "1.0.0",
   "description": "Template for the WebSocket remote control task",
-  "main": "index.js",
+  "main": "./src/index.ts",
   "type": "module",
   "engines": {
     "node": ">=16.15.0"
@@ -10,32 +10,23 @@
   "scripts": {
     "start": "npm run build && node ./dist/index.js",
     "build": "tsc",
-    "start:dev": "nodemon ./index.ts",
+    "start:dev": "nodemon ./src/index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Andrei Auchynnikau",
   "keywords": [],
   "license": "ISC",
   "devDependencies": {
+    "@types/node": "^17.0.42",
     "@types/ws": "^8.5.3",
-    "cross-env": "^7.0.3",
     "dotenv": "^16.0.1",
-    "eslint": "^8.17.0",
-    "nodemon": "^2.0.16",
-    "prettier": "^2.6.2",
+    "nodemon": "^1.3.3",
     "ts-node": "^10.8.1",
-    "typescript": "^4.7.3",
-    "uuid": "^8.3.2"
+    "typescript": "^4.7.3"
   },
   "dependencies": {
-    "@types/node": "^17.0.42",
-    "@types/supertest": "^2.0.12",
-    "@types/uuid": "^8.3.4",
-    "c8": "^7.11.3",
     "jimp": "^0.16.1",
     "robotjs": "^0.6.0",
-    "supertest": "^6.2.3",
-    "vitest": "^0.14.2",
     "ws": "^8.8.0"
   }
 }
